# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Shift Scheduler API",
    "description": "API for scheduling employee shifts."
  },
  "servers": [
    {
      "url": "http://shifts-processor.onrender.com"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "Welcome Endpoint",
        "description": "Returns a welcome message with an image of a Shih Tzu dog.",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "text/html": {
                "example": "<html>...</html>"
              }
            }
          }
        }
      }
    },
    "/schedule": {
      "post": {
        "summary": "Generate Shift Schedule",
        "description": "Generates a shift schedule based on the provided employees and days.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "employees": {
                    "type": "array",
                    "items": { "type": "string" },
                    "example": ["Alice", "Bob", "Charlie"]
                  },
                  "days": {
                    "type": "array",
                    "items": { "type": "string" },
                    "example": ["Monday", "Tuesday", "Wednesday"]
                  }
                },
                "required": ["employees", "days"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "CSV file with the schedule",
            "content": {
              "text/csv": {
                "example": "Day,Shift,Employee\nMonday,morning,Alice\n..."
              }
            }
          },
          "400": {
            "description": "Invalid input or no valid schedule found."
          }
        }
      }
    },
    "/pdf": {
      "get": {
        "summary": "PDF Password Removal Page",
        "description": "Returns an HTML form for uploading a password-protected PDF and removing its password.",
        "responses": {
          "200": {
            "description": "HTML upload form",
            "content": {
              "text/html": {
                "example": "<html>...</html>"
              }
            }
          }
        }
      }
    },
    "/pdf/remove-password": {
      "post": {
        "summary": "Remove PDF Password",
        "description": "Decrypts a password-protected PDF using the supplied (known) password and returns an unlocked copy. Does not guess or crack unknown passwords.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "The password-protected PDF file."
                  },
                  "password": {
                    "type": "string",
                    "description": "The password that opens the PDF."
                  }
                },
                "required": ["file", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The unlocked PDF as a file download.",
            "content": {
              "application/pdf": {}
            }
          },
          "400": {
            "description": "Missing file/password, wrong password, unencrypted PDF, or invalid PDF."
          }
        }
      }
    }
  }
}